+++ bin/tools.php
@@ -1,6 +1,6 @@
 <?php
 
-include_once __DIR__.'/../vendor/autoload.php';
+include_once __DIR__ . '/../vendor/autoload.php';
 
 function write_line(string $line)
 {