+++ src/AwsS3V3/AwsS3V3Adapter.php
@@ -14,6 +14,7 @@ use League\Flysystem\FilesystemAdapter;
 use League\Flysystem\FilesystemOperationFailed;
 use League\Flysystem\PathPrefixer;
 use League\Flysystem\StorageAttributes;
+use League\Flysystem\UnableToCheckDirectoryExistence;
 use League\Flysystem\UnableToCheckFileExistence;
 use League\Flysystem\UnableToCopyFile;
 use League\Flysystem\UnableToDeleteFile;
@@ -27,6 +28,8 @@ use League\MimeTypeDetection\FinfoMimeTypeDetector;
 use League\MimeTypeDetection\MimeTypeDetector;
 use Psr\Http\Message\StreamInterface;
 use Throwable;
+
+use function trim;
 
 class AwsS3V3Adapter implements FilesystemAdapter
 {
@@ -56,6 +59,16 @@ class AwsS3V3Adapter implements FilesystemAdapter
         'StorageClass',
         'Tagging',
         'WebsiteRedirectLocation',
+    ];
+    /**
+     * @var string[]
+     */
+    public const MUP_AVAILABLE_OPTIONS = [
+        'before_upload',
+        'concurrency',
+        'mup_threshold',
+        'params',
+        'part_size',
     ];
 
     /**
@@ -130,6 +143,19 @@ class AwsS3V3Adapter implements FilesystemAdapter
         }
     }
 
+    public function directoryExists(string $path): bool
+    {
+        try {
+            $prefix = $this->prefixer->prefixDirectoryPath($path);
+            $options = ['Bucket' => $this->bucket, 'Prefix' => $prefix, 'Delimiter' => '/'];
+            $command = $this->client->getCommand('ListObjects', $options);
+
+            return $this->client->execute($command)->hasKey('Contents');
+        } catch (Throwable $exception) {
+            throw UnableToCheckDirectoryExistence::forLocation($path, $exception);
+        }
+    }
+
     public function write(string $path, string $contents, Config $config): void
     {
         $this->upload($path, $contents, $config);
@@ -145,14 +171,14 @@ class AwsS3V3Adapter implements FilesystemAdapter
         $key = $this->prefixer->prefixPath($path);
         $options = $this->createOptionsFromConfig($config);
         $acl = $options['ACL'] ?? $this->determineAcl($config);
-        $shouldDetermineMimetype = $body !== '' && ! array_key_exists('ContentType', $options);
+        $shouldDetermineMimetype = $body !== '' && ! array_key_exists('ContentType', $options['params']);
 
         if ($shouldDetermineMimetype && $mimeType = $this->mimeTypeDetector->detectMimeType($key, $body)) {
-            $options['ContentType'] = $mimeType;
+            $options['params']['ContentType'] = $mimeType;
         }
 
         try {
-            $this->client->upload($this->bucket, $key, $body, $acl, ['params' => $options]);
+            $this->client->upload($this->bucket, $key, $body, $acl, $options);
         } catch (Throwable $exception) {
             throw UnableToWriteFile::atLocation($path, '', $exception);
         }
@@ -167,9 +193,17 @@ class AwsS3V3Adapter implements FilesystemAdapter
 
     private function createOptionsFromConfig(Config $config): array
     {
-        $options = [];
+        $options = ['params' => []];
 
         foreach (static::AVAILABLE_OPTIONS as $option) {
+            $value = $config->get($option, '__NOT_SET__');
+
+            if ($value !== '__NOT_SET__') {
+                $options['params'][$option] = $value;
+            }
+        }
+
+        foreach (static::MUP_AVAILABLE_OPTIONS as $option) {
             $value = $config->get($option, '__NOT_SET__');
 
             if ($value !== '__NOT_SET__') {
@@ -294,12 +328,7 @@ class AwsS3V3Adapter implements FilesystemAdapter
         $lastModified = $dateTime instanceof DateTimeResult ? $dateTime->getTimeStamp() : null;
 
         return new FileAttributes(
-            $path,
-            $fileSize,
-            null,
-            $lastModified,
-            $mimetype,
-            $this->extractExtraMetadata($metadata)
+            $path, $fileSize, null, $lastModified, $mimetype, $this->extractExtraMetadata($metadata)
         );
     }
 
@@ -406,7 +435,7 @@ class AwsS3V3Adapter implements FilesystemAdapter
                 $this->bucket,
                 $this->prefixer->prefixPath($destination),
                 $this->visibility->visibilityToAcl($visibility),
-                $this->createOptionsFromConfig($config)
+                $this->createOptionsFromConfig($config)['params']
             );
         } catch (Throwable $exception) {
             throw UnableToCopyFile::fromLocationTo($source, $destination, $exception);