+++ src/Filesystem.php
@@ -36,6 +36,18 @@ class Filesystem implements FilesystemOperator
         return $this->adapter->fileExists($this->pathNormalizer->normalizePath($location));
     }
 
+    public function directoryExists(string $location): bool
+    {
+        return $this->adapter->directoryExists($this->pathNormalizer->normalizePath($location));
+    }
+
+    public function has(string $location): bool
+    {
+        $path = $this->pathNormalizer->normalizePath($location);
+
+        return $this->adapter->fileExists($path) || $this->adapter->directoryExists($path);
+    }
+
     public function write(string $location, string $contents, array $config = []): void
     {
         $this->adapter->write(