+++ src/InMemory/InMemoryFilesystemAdapter.php
@@ -8,6 +8,7 @@ use League\Flysystem\Config;
 use League\Flysystem\DirectoryAttributes;
 use League\Flysystem\FileAttributes;
 use League\Flysystem\FilesystemAdapter;
+use League\Flysystem\FilesystemException;
 use League\Flysystem\UnableToCopyFile;
 use League\Flysystem\UnableToMoveFile;
 use League\Flysystem\UnableToReadFile;
@@ -16,6 +17,10 @@ use League\Flysystem\UnableToSetVisibility;
 use League\Flysystem\Visibility;
 use League\MimeTypeDetection\FinfoMimeTypeDetector;
 use League\MimeTypeDetection\MimeTypeDetector;
+
+use function array_keys;
+use function rtrim;
+use function strpos;
 
 class InMemoryFilesystemAdapter implements FilesystemAdapter
 {
@@ -105,6 +110,20 @@ class InMemoryFilesystemAdapter implements FilesystemAdapter
     {
         $filePath = rtrim($path, '/') . '/' . self::DUMMY_FILE_FOR_FORCED_LISTING_IN_FLYSYSTEM_TEST;
         $this->write($filePath, '', $config);
+    }
+
+    public function directoryExists(string $path): bool
+    {
+        $prefix = $this->preparePath($path);
+        $prefix = rtrim($prefix, '/') . '/';
+
+        foreach (array_keys($this->files) as $path) {
+            if (strpos($path, $prefix) === 0) {
+                return true;
+            }
+        }
+
+        return false;
     }
 
     public function setVisibility(string $path, string $visibility): void