+++ src/Local/LocalFilesystemAdapter.php
@@ -213,7 +213,7 @@ class LocalFilesystemAdapter implements FilesystemAdapter
             $permissions = octdec(substr(sprintf('%o', $fileInfo->getPerms()), -4));
             $visibility = $isDirectory ? $this->visibility->inverseForDirectory($permissions) : $this->visibility->inverseForFile($permissions);
 
-            yield $isDirectory ? new DirectoryAttributes($path, $visibility, $lastModified) : new FileAttributes(
+            yield $isDirectory ? new DirectoryAttributes(str_replace('\\', '/', $path), $visibility, $lastModified) : new FileAttributes(
                 str_replace('\\', '/', $path),
                 $fileInfo->getSize(),
                 $visibility,
@@ -302,6 +302,13 @@ class LocalFilesystemAdapter implements FilesystemAdapter
         $location = $this->prefixer->prefixPath($location);
 
         return is_file($location);
+    }
+
+    public function directoryExists(string $location): bool
+    {
+        $location = $this->prefixer->prefixPath($location);
+
+        return is_dir($location);
     }
 
     public function createDirectory(string $path, Config $config): void