+++ src/UnableToCheckFileExistence.php
@@ -4,16 +4,8 @@ declare(strict_types=1);
 
 namespace League\Flysystem;
 
-use RuntimeException;
-use Throwable;
-
-class UnableToCheckFileExistence extends RuntimeException implements FilesystemOperationFailed
+class UnableToCheckFileExistence extends UnableToCheckExistence
 {
-    public static function forLocation(string $path, Throwable $exception = null): UnableToCheckFileExistence
-    {
-        return new UnableToCheckFileExistence("Unable to check file existence for: ${path}", 0, $exception);
-    }
-
     public function operation(): string
     {
         return FilesystemOperationFailed::OPERATION_FILE_EXISTS;